requirements.txt
beautifulsoup4==4.10.0
PyYAML==6.0
selenium==4.1.2
